Add generic controller with CRUD endpoints and update service layer

+++ prateleira-inteligente-v2/src/main/java/br/com/lgbv/prateleira_inteligente_v2/services/GenericService.java
@@ -3,6 +3,7 @@ package br.com.lgbv.prateleira_inteligente_v2.services;
 import br.com.lgbv.prateleira_inteligente_v2.entities.BaseEntity;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.UUID;
@@ -13,6 +14,7 @@ public abstract class GenericService<E extends BaseEntity> implements IGenericSe
     protected final JpaRepository<E, UUID> repository;
 
     @Override
+    @Transactional
     public E save(E entity) {
         beforeSave(entity);
         E saved = repository.save(entity);
@@ -21,21 +23,25 @@ public abstract class GenericService<E extends BaseEntity> implements IGenericSe
     }
 
     @Override
+    @Transactional(readOnly = true)
     public E getById(UUID id) {
         return repository.findById(id).orElseThrow(() -> new RuntimeException("Entity not found"));
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<E> getAll() {
         return repository.findAll();
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<E> getAllByIds(List<UUID> ids) {
         return repository.findAllById(ids);
     }
 
     @Override
+    @Transactional
     public E update(UUID id,E entity) {
         beforeUpdate(entity);
         E saved = repository.save(entity);
@@ -44,8 +50,13 @@ public abstract class GenericService<E extends BaseEntity> implements IGenericSe
     }
 
     @Override
-    public void delete(UUID id) {
+    @Transactional
+    public boolean deleteById(UUID id) {
+        if (!repository.existsById(id)) {
+            return false;
+        }
         repository.deleteById(id);
+        return true;
     }
 
     protected void beforeSave(E entity) {
